fix: refactor to fix dep map overwrite problem
An issue potentially existed where if multiple versions of the same
dependency existed with different file paths then when they were
inserted into the dependency map the last entry would win.

This change makes it so that individual feed items are handled
in isolation therefore avoiding a shared map which risks overwrite.